- Backport new version to p7 branch

+++ lib/util/data_blob.c
@@ -18,7 +18,9 @@
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "includes.h"
+#include "replace.h"
+#include "attr.h"
+#include "data_blob.h"
 
 const DATA_BLOB data_blob_null = { NULL, 0 };
 
@@ -133,7 +135,7 @@ print the data_blob as hex string
 **/
 _PUBLIC_ char *data_blob_hex_string_lower(TALLOC_CTX *mem_ctx, const DATA_BLOB *blob)
 {
-	int i;
+	size_t i;
 	char *hex_string;
 
 	hex_string = talloc_array(mem_ctx, char, (blob->length*2)+1);
@@ -153,7 +155,7 @@ _PUBLIC_ char *data_blob_hex_string_lower(TALLOC_CTX *mem_ctx, const DATA_BLOB *
 
 _PUBLIC_ char *data_blob_hex_string_upper(TALLOC_CTX *mem_ctx, const DATA_BLOB *blob)
 {
-	int i;
+	size_t i;
 	char *hex_string;
 
 	hex_string = talloc_array(mem_ctx, char, (blob->length*2)+1);
